Use a podcast specific scaffold
